quic: don't destroy stream before app read data
In quic_quicly_on_stream_destroy do not free stream contex if
application has unread data.
Applications must now send rx events on data read.

Type: fix

+++ src/plugins/quic_quicly/quic_quicly_error.c
@@ -10,7 +10,7 @@
 u8 *
 quic_quicly_format_err (u8 *s, va_list *args)
 {
-  u64 code = va_arg (*args, u64);
+  int code = va_arg (*args, int);
   switch (code)
     {
     case 0:
